refactor : token_type, token 분리
