+++ Compiler/frontend/token/token.cpp
@@ -23,13 +23,9 @@ Token& Token::operator=(const Token& token) {
 	return *this;
 }
 
-Token::Token(const Token&& token) noexcept {
+//Token::Token(const Token&& token) noexcept {}
 
-}
-
-Token& Token::operator=(const Token&& src) noexcept {
-
-}
+//Token& Token::operator=(const Token&& src) noexcept {}
 
 Token::Token(const Token& copy) : type_(copy.type_), literal_(copy.literal_) {}
