+++ Compiler/frontend/token/token.h
@@ -16,8 +16,8 @@ public:
 
 	Token(const TokenType type, const std::string literal);
 	Token& operator=(const Token& token);
-	Token(const Token&& token) noexcept;
-	Token& operator=(const Token&& src) noexcept;
+	//Token(const Token&& token) noexcept;
+	//Token& operator=(const Token&& src) noexcept;
 	Token(const Token& copy);
 	~Token();
 };
